Add WrappedCommand for wrapping the DeleteCommand
